feat(matrix): Save messages to Store's buffer instead of explicitly sending them *now*
Fixes #1.

Basically, makes things magically work by taking advantage of all the goodness the SDK provides: namely, an `UnsentEventManager` and a `NetworkConnectivityReceiver`. And this all even happens under the hood~

Also, 'bit of a refactor: we now have a `MotoricaMechanicalMessage extends Message` class which represents our Matrix datagram. So, a bit more types and a bit less passing JSON around.